+++ src/main/kotlin/com/atlassian/performance/tools/jirasoftwareactions/CompatibleScrumBoardMemory.kt
@@ -3,6 +3,7 @@ package com.atlassian.performance.tools.jirasoftwareactions
 import com.atlassian.performance.tools.jiraactions.api.memories.Memory
 import com.atlassian.performance.tools.jirasoftwareactions.api.boards.ScrumBoard
 import com.atlassian.performance.tools.jirasoftwareactions.api.memories.AgileBoardIdMemory
+import com.atlassian.performance.tools.jirasoftwareactions.api.memories.BoardMemory
 
 /**
  * Translates the deprecated [AgileBoardIdMemory] into a [Memory] of [ScrumBoard]s.
@@ -11,10 +12,14 @@ import com.atlassian.performance.tools.jirasoftwareactions.api.memories.AgileBoa
  */
 internal class CompatibleScrumBoardMemory(
     private val boardIdMemory: AgileBoardIdMemory
-) : Memory<ScrumBoard> {
+) : BoardMemory<ScrumBoard> {
     override fun recall(): ScrumBoard? {
         val id = boardIdMemory.recall()
         return if (id != null) ScrumBoard(id) else null
+    }
+
+    override fun recall(filter: (ScrumBoard) -> Boolean): ScrumBoard? {
+        return recall()
     }
 
     override fun remember(memories: Collection<ScrumBoard>) {