+++ src/main/kotlin/com/atlassian/performance/tools/jirasoftwareactions/api/actions/BrowseBoardsAction.kt
@@ -3,22 +3,21 @@ package com.atlassian.performance.tools.jirasoftwareactions.api.actions
 import com.atlassian.performance.tools.jiraactions.api.BROWSE_BOARDS
 import com.atlassian.performance.tools.jiraactions.api.action.Action
 import com.atlassian.performance.tools.jiraactions.api.measure.ActionMeter
-import com.atlassian.performance.tools.jiraactions.api.memories.Memory
 import com.atlassian.performance.tools.jirasoftwareactions.CompatibleBoardMemory
 import com.atlassian.performance.tools.jirasoftwareactions.CompatibleScrumBoardMemory
 import com.atlassian.performance.tools.jirasoftwareactions.api.WebJiraSoftware
-import com.atlassian.performance.tools.jirasoftwareactions.api.boards.Board
 import com.atlassian.performance.tools.jirasoftwareactions.api.boards.AgileBoard
 import com.atlassian.performance.tools.jirasoftwareactions.api.boards.ScrumBoard
 import com.atlassian.performance.tools.jirasoftwareactions.api.memories.AgileBoardIdMemory
+import com.atlassian.performance.tools.jirasoftwareactions.api.memories.BoardMemory
 import net.jcip.annotations.NotThreadSafe
 
 @NotThreadSafe
 class BrowseBoardsAction(
     private val jiraSoftware: WebJiraSoftware,
     private val meter: ActionMeter,
-    private val boardsMemory: Memory<Board>,
-    private val scrumBoardsMemory: Memory<ScrumBoard>
+    private val boardsMemory: BoardMemory<AgileBoard>,
+    private val scrumBoardsMemory: BoardMemory<ScrumBoard>
 ) : Action {
 
     constructor(
@@ -29,8 +28,12 @@ class BrowseBoardsAction(
         jiraSoftware = jiraSoftware,
         meter = meter,
         boardsMemory = CompatibleBoardMemory(boardsMemory),
-        scrumBoardsMemory = object : Memory<ScrumBoard> {
+        scrumBoardsMemory = object : BoardMemory<ScrumBoard> {
             override fun recall(): ScrumBoard? {
+                return null
+            }
+
+            override fun recall(filter: (ScrumBoard) -> Boolean): ScrumBoard? {
                 return null
             }