+++ src/main/kotlin/com/atlassian/performance/tools/jirasoftwareactions/api/boards/Board.kt
@@ -2,6 +2,7 @@ package com.atlassian.performance.tools.jirasoftwareactions.api.boards
 
 interface Board {
     val id: String
-    val issuesOnBoard: Int?
+    var issuesOnBoard: Int?
+    @Deprecated(message = "Use ViewBoardAction.filter")
     fun isWorthVisiting(): Boolean
 }