+++ src/main/kotlin/com/atlassian/performance/tools/jirasoftwareactions/api/memories/AdaptiveBoardMemory.kt
@@ -1,26 +1,24 @@
 package com.atlassian.performance.tools.jirasoftwareactions.api.memories
 
 import com.atlassian.performance.tools.jiraactions.api.SeededRandom
-import com.atlassian.performance.tools.jiraactions.api.memories.Memory
 import com.atlassian.performance.tools.jirasoftwareactions.api.boards.Board
 import net.jcip.annotations.NotThreadSafe
 
 @NotThreadSafe
 class AdaptiveBoardMemory<T : Board>(
     private val random: SeededRandom
-) : Memory<T> {
+) : BoardMemory<T> {
     private val boards = mutableSetOf<T>()
 
     override fun recall(): T? {
         return random.pick(boards.toList())
     }
 
+    override fun recall(filter: (T) -> Boolean): T? {
+        return random.pick(boards.asSequence().filter(filter).toList())
+    }
+
     override fun remember(memories: Collection<T>) {
-        memories.forEach {
-            if (!boards.add(it)) {
-                boards.remove(it)
-                boards.add(it)
-            }
-        }
+        boards.addAll(memories)
     }
 }